Atualiza a configuração do projeto e refatora o servidor MCP
